Fix unshown claimed item
When an item had more than 1 quantity and some were claimed it would
show as unclaimed but there would be no copy in claimed. This would also
cause the category to show with nothing under it.

The handling of the default value is a bit odd but when I tried to
change the default to 0 there ended up being more changes vs. this one
place.