Make unlimited quantity items claimable
Changing the claim UI from the list view to be its own view. This is
done to support having claim notes. Claim notes are needed for the
ability to claim items that have unlimited quantity so you can say what
was claimed. Avoid duplicates when its possible to have them.

Refatored how the ItemsGrouper works. There was a lot of extra stuff
happening for getting the categories and duplicate efforts to determine
item claim state in the Item model.