List Item Grouper
Creating a replacement for the ItemGrouper service. The ItemGrouper is
hitting memory issues on Fly.io as it duplicates data and isn't the most
efficient. This new service will be more efficient as it does more in the
query and only collects 2 hashes for claimed and unclaimed items.

Not using the new service in this commit. Getting it ready to use and
tests in place.